+++ app.py
@@ -18,20 +18,6 @@ def index():
     return render_template('index.html')
 
 
-def gen(camera: gbv.Camera):
-    """Video streaming generator function."""
-    while True:
-        _, frame = camera.read()
-        yield (b'--frame\r\n'
-               b'Content-Type: image/jpeg\r\n\r\n' + cv2.imencode('.jpg', frame)[1].tobytes() + b'\r\n')
-
-
-@app.route('/video_feed')
-def video_feed():
-    """Video streaming route. Put this in the src attribute of an img tag."""
-    return Response(gen(cam),
-                    mimetype='multipart/x-mixed-replace; boundary=frame')
-
 
 if __name__ == '__main__':
     app.run(host='0.0.0.0', threaded=True)
